Move remaining tests out of tests folder
